Set up book and label actions in different files, also add the preserve data functions
