+++ methods/label.rb
@@ -1,11 +1,11 @@
 require_relative './item'
 
-class Label < Item
+class Label
   attr_accessor :title, :color
   attr_reader :id, :items
 
-  def initialize(title, color, items, id = Random.rand(1..10_000))
-    super(id)
+  def initialize(title, color, items, )
+    id = Random.rand(1..10_000)
     @title = title
     @color = color
     @items = []
